Split SteeringType into its own source file
