+++ lib/go-tc/enum/enum.go
@@ -124,48 +124,6 @@ func DSTypeCategoryFromString(s string) DSTypeCategory {
 const SigningAlgorithmURLSig = "url_sig"
 const SigningAlgorithmURISigning = "uri_signing"
 
-type SteeringType string
-
-const (
-	SteeringTypeOrder     SteeringType = "STEERING_ORDER"
-	SteeringTypeWeight    SteeringType = "STEERING_WEIGHT"
-	SteeringTypeGeoOrder  SteeringType = "STEERING_GEO_ORDER"
-	SteeringTypeGeoWeight SteeringType = "STEERING_GEO_WEIGHT"
-	SteeringTypeInvalid   SteeringType = ""
-)
-
-func SteeringTypeFromString(s string) SteeringType {
-	s = strings.ToLower(strings.Replace(s, "_", "", -1))
-	switch s {
-	case "steeringorder":
-		return SteeringTypeOrder
-	case "steeringweight":
-		return SteeringTypeWeight
-	case "steeringgeoorder":
-		return SteeringTypeGeoOrder
-	case "steeringgeoweight":
-		return SteeringTypeGeoWeight
-	default:
-		return SteeringTypeInvalid
-	}
-}
-
-// String returns a string representation of this steering type.
-func (t SteeringType) String() string {
-	switch t {
-	case SteeringTypeOrder:
-		fallthrough
-	case SteeringTypeWeight:
-		fallthrough
-	case SteeringTypeGeoOrder:
-		fallthrough
-	case SteeringTypeGeoWeight:
-		return string(t)
-	default:
-		return "INVALID"
-	}
-}
-
 type FederationResolverType string
 
 const (
